Adding the ability to use formulae in unit definitions.
There are some caveats to doing this that I would like to find a
way to rectify.  Currently formulae used to define a unit need to
have dimensionless inputs and dimensions should be introduced in
the formula body, usually by multiplying with a previously defined
unit.  This works, but I would like to make the inputs to the formula
reflect the units being specified.  This way the formula definitions
can be used in a more general capacity rather than just for unit definition.