Refactored the code I've added to be better integrated and less hacky.
Units now have slots that govern how they convert-to and convert-from base units.
Typically they hold the symbols  and , which is the default behavior that
unit-formulas has always used.  By specifying a unit definition that uses the
symbol formula as the first item in a list, the rest of the list will be passed
as arguments to make-instance 'unit allowing for formulae to be passed as
the convert-to and convert-from functions.  See the additions to the documentation
as well as examples used in the oeno.units library.